_/deep-reseach/requirements.txt
tavily-python
beautifulsoup4
httpx
langchain
langchain-openai
